requirements.txt
openai-agents
fastapi
uvicorn
sqlmodel
aiosqlite
requests
python-multipart
python-dotenv

fastapi
uvicorn
sqlmodel
aiosqlite
requests
